[feat] Added Bachelor of Creative Tech & AI content
The first steps have been made.
Is accessible at `mct.be/international/bachelor-creative-tech-ai` where the whole programme can be found.

The content of the modules hasn't been added yet. This will be something to do the coming days, together with the other colleagues.

Issue noticed: When accessing a specific module, the `< Het volledige programma` should be translated to English if needed.
We should add the `english: true` label to the courses if needed.

Also, some courses are exactly the same as MCT. We should see if we can make it better in the new system with SvelteKit ...